LayoutTests/platform/blackberry/editing/selection/after-line-break-expected.txt

EDITING DELEGATE: shouldChangeSelectedDOMRange:(null) toDOMRange:range from 275 of #text > P > BODY > HTML > #document to 275 of #text > P > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 275 of #text > P > BODY > HTML > #document to 275 of #text > P > BODY > HTML > #document toDOMRange:range from 0 of P > BODY > HTML > #document to 0 of P > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
Test for rdar://problem/5973313 REGRESSION (r32508): Down arrow doesn't change caret with non-user entered newlines in textareas.

To test manually, click to the right of the first word on the first line to place the caret at the end of the line, then press the down arrow. The caret should move to the next (empty) line. Then click somewhere on that line. The caret should not stay there.

Lorem

ipsum
